Drop insider docs and switch to SCM versioning
