+++ .gitattributes
@@ -1,1 +1,1 @@
-femto/fe/_version.py export-subst
+.git_archival.txt  export-subst
