+++ .github/workflows/ci.yaml
@@ -24,10 +24,11 @@ jobs:
         make env
         make lint
         make test
-        make docs-build
+        make docs
 
     - name: CodeCov
-      uses: codecov/codecov-action@v3.1.1
+      uses: codecov/codecov-action@v4.1.1
       with:
         file: ./coverage.xml
         flags: unittests
+        token: ${{ secrets.CODECOV_TOKEN }}
