+++ Makefile
@@ -1,10 +1,10 @@
 PACKAGE_NAME := femto
 
-CONDA_ENV_RUN=conda run --no-capture-output --name femto
+CONDA_ENV_RUN=conda run --no-capture-output --name $(PACKAGE_NAME)
 
 TEST_ARGS := -v --cov=$(PACKAGE_NAME) --cov-report=term --cov-report=xml --junitxml=unit.xml --color=yes
 
-.PHONY: env lint format test docs-build docs-deploy docs-insiders
+.PHONY: env lint format test docs docs-deploy docs-insiders
 
 env:
 	mamba create     --name $(PACKAGE_NAME)
@@ -18,11 +18,12 @@ lint:
 
 format:
 	$(CONDA_ENV_RUN) ruff format $(PACKAGE_NAME)
+	$(CONDA_ENV_RUN) ruff check --fix --select I $(PACKAGE_NAME)
 
 test:
 	$(CONDA_ENV_RUN) pytest -v $(TEST_ARGS) $(PACKAGE_NAME)/*/tests/
 
-docs-build:
+docs:
 	$(CONDA_ENV_RUN) mkdocs build
 
 docs-deploy:
@@ -30,7 +31,3 @@ ifndef VERSION
 	$(error VERSION is not set)
 endif
 	$(CONDA_ENV_RUN) mike deploy --push --update-aliases $(VERSION)
-
-docs-insiders:
-	$(CONDA_ENV_RUN) pip install git+https://$(INSIDER_DOCS_TOKEN)@github.com/SimonBoothroyd/mkdocstrings-python.git \
-                    			 git+https://$(INSIDER_DOCS_TOKEN)@github.com/SimonBoothroyd/griffe-pydantic.git@fix-inheritence-static
