+++ devtools/envs/base.yaml
@@ -55,6 +55,7 @@ dependencies:
   - mkdocs-gen-files
   - mkdocs-literate-nav
   - mkdocstrings
-  - mkdocstrings-python
+  - mkdocstrings-python >=1.10.8
+  - griffe-pydantic
   - black
   - mike
