+++ mkdocs.yml
@@ -103,13 +103,12 @@ plugins:
           - http://docs.openmm.org/latest/api-python/objects.inv
           - https://www.mdtraj.org/1.9.6/objects.inv
         options:
-          # extensions: [ griffe_pydantic ]
+          extensions: [ griffe_pydantic ]
           docstring_options:
             ignore_init_summary: true
             returns_multiple_items: false
             returns_named_value: false
           docstring_section_style: list
-          filters: ["!^_", "!^model_dump_yaml$", "!^model_config$"]
           heading_level: 1
           inherited_members: true
           merge_init_into_class: true
@@ -119,7 +118,6 @@ plugins:
           show_signature_annotations: true
           show_symbol_type_heading: true
           show_symbol_type_toc: true
-          show_if_no_docstring: false
           signature_crossrefs: true
           summary: true
           members_order: source
