+++ pyproject.toml
@@ -1,5 +1,5 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel", "versioneer"]
+requires = ["setuptools>=61.0", "setuptools_scm>=8", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
@@ -17,23 +17,15 @@ femto = "femto.fe._cli:main"
 femto = "femto.fe._cli:main"
 
 [tool.setuptools]
-zip-safe = false
 include-package-data = true
 
 [tool.setuptools.packages.find]
-namespaces = true
-where = ["."]
+include = ["femto*"]
 
 [tool.setuptools.package-data]
 "femto.fe.testing" = ["data/**/*"]
 
-[tool.versioneer]
-VCS = "git"
-style = "pep440"
-versionfile_source = "femto/fe/_version.py"
-versionfile_build = "femto/fe/_version.py"
-tag_prefix = ""
-parentdir_prefix = "femto-"
+[tool.setuptools_scm]
 
 [tool.ruff.lint]
 ignore = ["C901"]
@@ -41,7 +33,7 @@ select = ["B","C","E","F","W","B9"]
 ignore-init-module-imports = true
 
 [tool.coverage.run]
-omit = ["**/tests/*", "**/_version.py"]
+omit = ["**/tests/*"]
 
 [tool.coverage.report]
 exclude_lines = [
@@ -54,7 +46,7 @@ exclude_lines = [
 ]
 
 [tool.vulture]
-exclude = ["versioneer.py", "_version.py", "cli.py"]
+exclude = ["cli.py"]
 ignore_names = ["process", "expand", "setup", "mode", "mock_get_summary", "cls"]
 min_confidence = 62
 paths = ["femto"]
