--- setup.py
@@ -1,4 +1,0 @@
-import versioneer
-from setuptools import setup
-
-setup(version=versioneer.get_version(), cmdclass=versioneer.get_cmdclass())
